feat!(measurements): use uint64_t as saveThreshold datatype

+++ examples/wordCount/lines2words/lines2words.cpp
@@ -97,7 +97,7 @@ int main(int argc, const char *argv[]) {
   std::string pubTopic = producerConfig["topic"].as<std::string>();
   std::string subTopic = consumerConfig["topic"].as<std::string>();
   int publishInterval = producerConfig["publishInterval"].as<int>();
-  int saveThreshold = measurementConfig["saveThreshold"].as<int>();
+  uint64_t saveThreshold = measurementConfig["saveThreshold"].as<uint64_t>();
 
   ::signal(SIGINT, signalCallbackHandler);
   measurementHandler = new iceflow::Measurement(measurementFileName, nodePrefix,
